Move poc into library & added temp way of providing localizations
This is already taking shape. The host builder extension can be used to install the desired localization provider you want to use. This is now in a place where I can create some goodies like xmlns definitions and then start working on a caching layer. I do need to create a proper way of managing the config within the library but baby steps...