App resources now loaded from configuration
Started work on the config layer for the library, you can define what your default culture is and it will be the first thing loaded in. Had a few points where I will need to think harder down the line such as how we determine what the default culture is, if the user doesnt set it (query device?). Added internals visible to, will need to watch that nick chapsas video to get the juice as I cant remember what he did to make that stuff awesome. Also lots of static singletons creeping in so need to be careful!