Added a service locator, unfortunately
I was having a play to see if I could do this lib with zero service locators, looks like the converters and any external markup extensions would be more difficult to write if I dont provide this workaround. I think there would be an easy anaylzer to make this not widely used, just where it needs to be (aka a converter). Feels more like a maui problem though :)