UriExtension tests
Testing this turned out to be a huge hassle. The TryAppend method was a bit pep and didn't behave at all how I would of expected it to! I've refactored it to validate the uri string on the way out since creating an absolute uri will never throw an exception regardless of how bad the input string is!